Remove summator from name
