+++ zaitsev.vladimir/S5/summator.hpp
@@ -1,0 +1,14 @@
+#ifndef SUMMATOR_HPP
+#define SUMMATOR_HPP
+#include <string>
+
+namespace zaitsev
+{
+  struct summator
+  {
+    void operator()(const std::pair< const long long, std::string >& key_value);
+    long long key_sum_ = 0;
+    std::string val_sum_ = "";
+  };
+}
+#endif
